Functions

Functions

Index

GigaSOM.Som

GigaSOM.daFrame

GigaSOM.classFrequencies

GigaSOM.cleanNames!

GigaSOM.convertTrainingData

GigaSOM.createDaFrame

GigaSOM.distMatrix

GigaSOM.doEpoch

GigaSOM.findBmu

GigaSOM.gaussianKernel

GigaSOM.getMarkers

GigaSOM.gridRectangular

GigaSOM.initGigaSOM

GigaSOM.makeClassFreqs

GigaSOM.makePopulation

GigaSOM.mapToGigaSOM

GigaSOM.normTrainData

GigaSOM.normTrainData

GigaSOM.readFlowset

GigaSOM.trainGigaSOM

GigaSOM.transformData

GigaSOM.visual

Input

GigaSOM.readFlowset — Method.

readFlowset(filenames)

Create a dictionary with filenames as keys and daFrame as values

Arguments:

filenames: Array of type string

source

Process

GigaSOM.cleanNames! — Method.

cleanNames!(mydata)

Replaces problematic characters in column names. Checks if the column name contains a '-' and transforms it to and '_' and it checks if the name starts with a number.

Arguments:

mydata: dict fcsRaw or array of string

source

GigaSOM.createDaFrame — Method.

createDaFrame(fcsRaw, md, panel)

Creates a daFrame of type struct. Read in the fcs raw, add sample id, subset the columns and transform

Arguments:

fcsRaw: raw FCS data

md: Metadata table

panel: Panel table with a column for Lineage Markers and one for Functional Markers

source

GigaSOM.getMarkers — Method.

getMarkers(panel)

Returns the lineageMarkers and functionalMarkers on a given panel

Arguments:

panel: Panel table with a column for Lineage Markers and one for Functional Markers

source

GigaSOM.transformData — Function.

transformData(flowframe, method = "asinh", cofactor = 5)

Tansforms FCS data. Currently only asinh

Arguments:

flowframe: Flowframe containing daFrame per sample

method: transformation method

cofactor: Cofactor for transformation

source

Core

GigaSOM.initGigaSOM — Function.

    initGigaSOM(train, xdim, ydim = xdim;  norm = :none, toroidal = false)

Initialises a SOM.

Arguments:

train: training data

xdim, ydim: geometry of the SOM If DataFrame, the column names will be used as attribute names. Codebook vectors will be sampled from the training data.

norm: optional normalisation; one of :minmax, :zscore or :none

toroidal: optional flag; if true, the SOM is toroidal.

source

GigaSOM.mapToGigaSOM — Method.

mapToGigaSOM(som::Som, data)

Return a DataFrame with X-, Y-indices and index of winner neuron for every row in data.

Arguments

som: a trained SOM

data: Array or DataFrame with training data.

Data must have the same number of dimensions as the training dataset and will be normalised with the same parameters.

source

GigaSOM.trainGigaSOM — Method.

trainGigaSOM(som::Som, train::Any, kernelFun = gaussianKernel,
                    r = 0.0, epochs = 10)

Arguments:

som: object of type Som with an initialised som

train: training data

kernel::function: optional distance kernel; one of (bubbleKernel, gaussianKernel) default is gaussianKernel

r: optional training radius. If r is not specified, it defaults to √(xdim^2 + ydim^2) / 2

Training data must be convertable to Array{Float34,2} with convert(). Training samples are row-wise; one sample per row. An alternative kernel function can be provided to modify the distance-dependent training. The function must fit to the signature fun(x, r) where x is an arbitrary distance and r is a parameter controlling the function and the return value is between 0.0 and 1.0.

source

GigaSOM.doEpoch — Method.

doEpoch(x::Array{Float64}, codes::Array{Float64}, dm::Array{Float64},
        kernelFun::Function, r::Number, toroidal::Bool)

vectors and the adjustment in radius after each epoch.

Arguments:

x: training Data

codes: Codebook

dm: distance matrix of all neurons of the SOM

kernelFun: distance kernel function of type fun(x, r)

r: training radius

toroidal: if true, the SOM is toroidal.

source

Structs

GigaSOM.daFrame — Type.

daFrame

Structure to hold FCS files.

Fields:

fcstable::Dict: Key: Filename, Value: FCS data table

md::DataFrame: Metadata containing the filenames, and experiment conditions

panel::DataFrame: defines which markers to use (lineage, functional)

source

GigaSOM.Som — Type.

Som

Structure to hold all data of a trained SOM.

Fields:

codes::Array{Float64,2}: 2D-array of codebook vectors. One vector per row

colNames::Array{String,1}: names of the attribute with which the SOM is trained

normParams::DataFrame: normalisation parameters for each column of training data. Column headers corresponds with colNames.

norm::Symbol: normalisation type; one of :none, :minmax, :zscore

xdim::Int: number of neurons in x-direction

ydim::Int: number of neurons in y-direction

numCodes::Int: total number of neurons

grid::Array{Float64,2}: 2D-array of coordinates of neurons on the map (2 columns (x,y)] for rectangular and hexagonal maps 3 columns (x,y,z) for spherical maps)

indices::DataFrame: X-, Y-indices of the neurons

topol::Symbol: topology of the SOM; one of :rectangular, :hexagonal, :spherical

toroidal::Bool: if true, the SOM is toroidal (has no edges)

population::Array{Int,1}: 1D-array of numbers of training samples mapped to each neuron.

source

Satellites

GigaSOM.classFrequencies — Method.

classFrequencies(som::Som, data, classes)

Return a DataFrame with class frequencies for all neurons.

Arguments:

som: a trained SOM

data: data with row-wise samples and class information in each row

classes: Name of column with class information.

Data must have the same number of dimensions as the training dataset. The column with class labels is given as classes (name or index). Returned DataFrame has the columns:

X-, Y-indices and index: of winner neuron for every row in data

population: number of samples mapped to the neuron

frequencies: one column for each class label.

source

GigaSOM.convertTrainingData — Method.

convertTrainingData(data)

Converts the training data to an Array of type Float64.

Arguments:

data: Data to be converted

source

GigaSOM.distMatrix — Method.

distMatrix(grid::Array, toroidal::Bool)

Return the distance matrix for a non-toroidal or toroidal SOM.

Arguments

grid: coordinates of all neurons as generated by one of the grid-functions

with x-coordinates in 1st column and y-coordinates in 2nd column.

toroidal: true for a toroidal SOM.

source

GigaSOM.findBmu — Method.

findBmu(codes, sample)

Find the best matching unit for a given vector, row_t, in the SOM

Returns: A (bmu, bmuIdx) tuple where bmu is the high-dimensional Best Matching Unit and bmuIdx is the index of this vector in the SOM

Arguments

codes: 2D-array of codebook vectors. One vector per row

sample: row in dataset / trainingsset

source

GigaSOM.gaussianKernel — Method.

gaussianKernel(x::Float64, r::Float64)::Float64

Return Gaussian(x) for μ=0.0 and σ = r/3. (a value of σ = r/3 makes the training results comparable between different kernels for same values of r).

Arguments

source

GigaSOM.gridRectangular — Method.

gridRectangular(xdim, ydim)

Create coordinates of all neurons on a rectangular SOM.

The return-value is an array of size (Number-of-neurons, 2) with x- and y- coordinates of the neurons in the first and second column respectively. The distance between neighbours is 1.0. The point of origin is bottom-left. The first neuron sits at (0,0).

Arguments

xdim: number of neurons in x-direction

ydim: number of neurons in y-direction

source

GigaSOM.makeClassFreqs — Method.

makeClassFreqs(som, vis, classes)

Return a DataFrame with class frequencies for all neurons.

Arguments

som: a trained SOM

vis: index of the winner neuron for each training pattern in x

classes: name of column with class information

source

GigaSOM.makePopulation — Method.

makePopulation(numCodes, vis)

Return a vector of neuron populations.

Arguments

numCodes: total number of neurons

vis: index of the winner neuron for each training pattern in x

source

GigaSOM.normTrainData — Method.

normTrainData(x::DataFrame, normParams::DataFrame)

Normalise every column of training data with the params.

Arguments

x: DataFrame with training Data

normParams: Shift and scale parameters for each attribute column.

source

GigaSOM.normTrainData — Method.

normTrainData(train::DataFrame, norm::Symbol)

Normalise every column of training data.

Arguments

train: DataFrame with training Data

norm: type of normalisation; one of minmax, zscore, none

source

GigaSOM.visual — Method.

visual(codes, x)

Return the index of the winner neuron for each training pattern in x (row-wise).

Arguments

codes: Codebook

x: training Data

source

Visualisation