Functions

Data structures

GigaSOM.LoadedDataInfo — Type

LoadedDataInfo

The basic structure for working with loaded data, distributed amongst workers. In completeness, it represents a dataset as such:

val is the "value name" under which the data are saved in processes. E.g. val=:foo means that there is a variable foo on each process holding a part of the matrix.

workers is a list of workers (in correct order!) that hold the data (similar to DArray.pids)

GigaSOM.Som — Type

Som

Structure to hold all data of a trained SOM.

Fields:

codes::Array{Float64,2}: 2D-array of codebook vectors. One vector per row

colNames::Array{String,1}: names of the attribute with which the SOM is trained

xdim::Int: number of neurons in x-direction

ydim::Int: number of neurons in y-direction

numCodes::Int: total number of neurons

grid::Array{Float64,2}: 2D-array of coordinates of neurons on the map (2 columns (x,y)] for rectangular and hexagonal maps 3 columns (x,y,z) for spherical maps)

Data loading and preparation

GigaSOM.distributeFCSFileVector — Function

distributeFCSFileVector(name::Symbol, fns::Vector{String}, pids=workers())::LoadedDataInfo

Distribute a vector of integers among the workers that describes which file from fns the cell comes from. Useful for producing per-file statistics. The vector is saved on workers specified by pids as a distributed variable name.

GigaSOM.distributeFileVector — Function

distributeFileVector(name::Symbol, sizes::Vector{Int}, slices::Vector{Tuple{Int,Int,Int,Int}}, pids=workers())::LoadedDataInfo

Generalized version of distributeFCSFileVector that produces the integer vector from any sizes and slices.

GigaSOM.getFCSSize — Method

getFCSSize(offsets, params)::Tuple{Int,Int}

Convert the offsets and keywords from an FCS file to cell and parameter count, respectively.

GigaSOM.loadFCS — Method

loadFCS(fn::String; applyCompensation::Bool=true)::Tuple{Dict{String,String}, Matrix{Float64}}

Read a FCS file. Return a tuple that contains in order:

dictionary of the keywords contained in the file

raw column names

prettified and annotated column names

raw data matrix

If applyCompensation is set, the function parses and retrieves a spillover matrix (if any valid keyword in the FCS is found that would contain it) and applies it to compensate the data.

GigaSOM.loadFCSHeader — Method

loadFCSHeader(fn::String)::Tuple{Vector{Int}, Dict{String,String}}

Efficiently extract data offsets and keyword dictionary from an FCS file.

GigaSOM.loadFCSSet — Function

loadFCSSet(name::Symbol, fns::Vector{String}, pids=workers(); applyCompensation=true, postLoad=(d,i)->d)::LoadedDataInfo

This runs the FCS loading machinery in a distributed way, so that the files fns (with full path) are sliced into equal parts and saved as a distributed variable name on workers specified by pids.

applyCompensation is passed to loadFCS function.

See slicesof for description of the slicing.

postLoad is applied to the loaded FCS file data (and the index) – use this function to e.g. filter out certain columns right on loading, using selectFCSColumns.

The loaded dataset can be manipulated by the distributed functions, e.g.

dselect for removing columns

dscale for normalization

dtransform_asinh (and others) for transformation

etc.

GigaSOM.loadFCSSizes — Method

loadFCSSizes(fns::Vector{String})

Load cell counts in many FCS files at once. Useful as input for slicesof.

GigaSOM.selectFCSColumns — Method

selectFCSColumns(selectColnames::Vector{String})

Return a function useful with loadFCSSet, which loads only the specified (prettified) column names from the FCS files. Use getMetaData, getMarkerNames and cleanNames! to retrieve the usable column names for a FCS.

GigaSOM.cleanNames! — Method

cleanNames!(mydata::Vector{String})

Replaces problematic characters in column names, avoids duplicate names, and prefixes an '_' if the name starts with a number.

Arguments:

mydata: vector of names (gets modified)

GigaSOM.compensate! — Method

compensate!(data::Matrix{Float64}, spillover::Matrix{Float64}, cols::Vector{Int})

Apply a compensation matrix in spillover (the individual columns of which describe, in order, the spillover of cols in data) to the matrix data in-place.

GigaSOM.getMarkerNames — Method

getMarkerNames(meta::DataFrame)::Tuple{Vector{String}, Vector{String}}

Extract suitable raw names (useful for selecting columns) and pretty readable names (useful for humans) from FCS file metadata.

GigaSOM.getMetaData — Method

getMetaData(f)

Collect the meta data information in a more user friendly format.

Arguments:

f: input structure with .params and .data fields

GigaSOM.getSpillover — Method

getSpillover(params::Dict{String, String})::Union{Tuple{Vector{String},Matrix{Float64}}, Nothing}

Get a spillover matrix from FCS params. Returns a pair with description of columns to be applied, and with the actual spillover matrix. Returns nothing in case spillover is not present.

GigaSOM.parseSpillover — Method

parseSpillover(str::String)::Union{Tuple{Vector{String},Matrix{Float64}}, Nothing}

Parses the spillover matrix from the string from FCS parameter value.

GigaSOM.collectSlice — Method

collectSlice(loadVec, (startFile, startOff, finalFile, finalOff)::Tuple{Int,Int,Int,Int})::Vector

Alternative of vcollectSlice for 1D vectors.

GigaSOM.slicesof — Method

slicesof(lengths::Vector{Int}, slices::Int)::Vector{Tuple{Int,Int,Int,Int}}

Given a list of lengths of input arrays, compute a slicing into a specified amount of equally-sized slices.

The output is a vector of 4-tuples where each specifies how to create one slice. The i-th tuple field contains, in order:

the index of input array at which the i-th slice begins

first element of the i-th slice in that input array

the index of input array with the last element of the i-th slice

the index of the last element of the i-th slice in that array

GigaSOM.vcollectSlice — Method

vcollectSlice(loadMtx, (startFile, startOff, finalFile, finalOff)::Tuple{Int,Int,Int,Int})::Matrix

Given a method to obtain matrix content (loadMtx), reconstruct a slice from the information generated by slicesof.

This function is specialized for reconstructing matrices and arrays, where the "element counts" split by slicesof are in fact matrix rows. The function is therefore named vcollect (the slicing and concatenation is vertical).

The actual data content and loading method is abstracted out – function loadMtx gets the index of the input part that it is required to fetch (e.g. index of one FCS file), and is expected to return that input part as a whole matrix. vcollectSlice correctly calls this function as required and extracts relevant portions of the matrices, so that at the end the whole slice can be pasted together.

Example:

# get a list of files
filenames=["a.fcs", "b.fcs"]
# get descriptions of 5 equally sized parts of the data
slices = slicesof(loadFCSSizes(filenames), 5)

# reconstruct first 3 columns of the first slice
mySlice = vcollectSlice(
    i -> last(loadFCS(slices[i]))[:,1:3],
    slices[1])
# (note: function loadFCS returns 4 items, the matrix is the last one)

GigaSOM.catmapbuckets — Method

catmapbuckets(fn, a::Array, nbuckets::Int, buckets::Vector{Int}; bucketdim::Int=1)

Same as mapbuckets, except concatenates the bucketing results in the bucketing dimension, thus creating a slightly neater matrix. slicedims is therefore fixed to bucketdim.

GigaSOM.collect_extrema — Method

collect_extrema(ex1, ex2)

Helper for collecting the minimums and maximums of the data. ex1, ex2 are arrays of pairs (min,max), this function combines the arrays element-wise and finds combined minima and maxima.

GigaSOM.combine_stats — Method

combine_stats((s1, sqs1, n1), (s2, sqs2, n2))

Helper for dstat-style functions that just adds up elements in triplets of vectors.

GigaSOM.dapply_cols — Method

dapply_cols(dInfo::LoadedDataInfo, fn, columns::Vector{Int})

Apply a function fn over columns of a distributed dataset.

fn gets 2 parameters:

a data vector for (the whole column saved at one worker)

index of the column in the columns array (i.e. a number from 1:length(columns))

GigaSOM.dapply_rows — Method

dapply_rows(dInfo::LoadedDataInfo, fn)

Apply a function fn over rows of a distributed dataset.

fn gets a single vector parameter for each row to transform.

GigaSOM.dcopy — Method

dcopy(dInfo::LoadedDataInfo, newName::Symbol)

Clone the dataset and store it under a new distributed name newName.

GigaSOM.dcount — Method

dcount(ncats::Int, dInfo::LoadedDataInfo)::Vector{Int}

Count the numbers of integer vector values stored in dInfo; assuming the values are in range 1–ncats.

GigaSOM.dcount_buckets — Method

dcount_buckets(ncats::Int, dInfo::LoadedDataInfo, nbuckets::Int, buckets::LoadedDataInfo)::Matrix{Int}

Same as dcount, but counts the items in dInfo bucketed by buckets to produce a matrix of counts, with ncats rows and nbuckets columns.

Useful with distributeFCSFileVector to determine cluster distribution within files.

GigaSOM.dmedian — Method

dmedian(dInfo::LoadedDataInfo, columns::Vector{Int})

Compute a median in a distributed fashion, avoiding data transfer and memory capacity that is required to compute the median in the classical way by sorting. All data must be finite and defined. If the median is just between 2 values, the lower one is chosen.

The algorithm is approximative, searching for a good median by halving interval and counting how many values are below the threshold. iters can be increased to improve precision, each value adds roughly 1 bit to the precision. The default value is 20, which corresponds to precision 10e-6 times the data range.

GigaSOM.dmedian_buckets — Method

dmedian_buckets(dInfo::LoadedDataInfo, nbuckets::Int, buckets::LoadedDataInfo, columns::Vector{Int}; iters=20)

A version of dmedian that works with the bucketing information (i.e. clusters) from nbuckets and buckets.

GigaSOM.dscale — Method

dscale(dInfo::LoadedDataInfo, columns::Vector{Int})

Scale the columns in the dataset to have mean 0 and sdev 1.

Prevents creation of NaNs by avoiding division by zero sdevs.

GigaSOM.dselect — Function

dselect(dInfo::LoadedDataInfo, columns::Vector{Int}; tgt=dInfo.val)

Reduce dataset to selected columns, optionally save it under a different name.

GigaSOM.dselect — Function

function dselect(dInfo::LoadedDataInfo,
    currentColnames::Vector{String}, selectColnames::Vector{String};
    tgt=dInfo.val)::LoadedDataInfo

Convenience overload of dselect that works with column names.

GigaSOM.dstat — Method

dstat(dInfo::LoadedDataInfo, columns::Vector{Int})

Compute mean and standard deviation of the columns in dataset. Returns a tuple with a vector of means in columns, and a vector of corresponding sdevs.

GigaSOM.dstat_buckets — Method

dstat_buckets(dInfo::LoadedDataInfo, nbuckets::Int, buckets::LoadedDataInfo, columns::Vector{Int})

A version of dstat that works with bucketing information (e.g. clusters).

GigaSOM.dtransform_asinh — Function

dtransform_asinh(dInfo::LoadedDataInfo, columns::Vector{Int}, cofactor=5)

Transform columns of the dataset by asinh transformation with cofactor.

GigaSOM.mapbuckets — Method

mapbuckets(fn, a::Array, nbuckets::Int, buckets::Vector{Int}; bucketdim::Int=1, slicedims=bucketdim)

Apply the function fn over array a so that it processes the data by buckets defined by buckets (that contains integers in range 1:nbuckets).

The buckets are sliced out in dimension specified by bucketdim.

GigaSOM.update_extrema — Method

update_extrema(counts, target, lim, mid)

Helper for distributed median computation – returns updated extrema in lims depending on whether the item count in counts of values less than mids is less or higher than targets.

SOM training

GigaSOM.distributedEpoch — Method

distributedEpoch(dInfo::LoadedDataInfo, codes::Matrix{Float64}, tree)

Execute the doEpoch in parallel on workers described by dInfo and collect the results. Returns pair of numerator and denominator matrices.

GigaSOM.doEpoch — Method

doEpoch(x::Array{Float64, 2}, codes::Array{Float64, 2}, tree)

vectors and the adjustment in radius after each epoch.

Arguments:

x: training Data

codes: Codebook

tree: knn-compatible tree built upon the codes

GigaSOM.initGigaSOM — Function

initGigaSOM(trainInfo::LoadedDataInfo,
            xdim::Int64, ydim :: Int64 = xdim)

initGigaSOM overload for working with distributed-style LoadedDataInfo data. The rest of arguments is the same as in initGigaSOM.

Note that this function only uses the data saved on the first worker for initialization, and the init work is actually done on that worker to avoid unnecessary data copying.

GigaSOM.initGigaSOM — Function

initGigaSOM(train, xdim, ydim = xdim)

Initializes a SOM by random selection from the training data.

Arguments:

train: codeBook vector as random input matrix from random workers

xdim, ydim: geometry of the SOM

GigaSOM.mapToGigaSOM — Method

mapToGigaSOM(som::Som, data;
             knnTreeFun = BruteTree,
             metric = Euclidean())

Overload of mapToGigaSOM for simple DataFrames and matrices. This slices the data using DistributedArrays, sends them the workers, and runs normal mapToGigaSOM. Data is undistributed after the computation.

GigaSOM.mapToGigaSOM — Method

mapToGigaSOM(som::Som, dInfo::LoadedDataInfo;
    knnTreeFun = BruteTree, metric = Euclidean(),
    output::Symbol=tmpSym(dInfo)::LoadedDataInfo

Compute the index of the BMU for each row of the input data.

Arguments

som: a trained SOM

dInfo: LoadedDataInfo that describes the loaded and distributed data

knnTreeFun: Constructor of the KNN-tree (e.g. from NearestNeighbors package)

metric: Passed as metric argument to the KNN-tree constructor

output: Symbol to save the result, defaults to tmpSym(dInfo)

Data must have the same number of dimensions as the training dataset and will be normalised with the same parameters.

GigaSOM.scaleEpochTime — Method

scaleEpochTime(iteration::Int64, epochs::Int64)

Convert iteration ID and epoch number to relative time in training.

GigaSOM.trainGigaSOM — Method

trainGigaSOM(som::Som, train;
             kernelFun::Function = gaussianKernel,
             metric = Euclidean(),
             somDistFun = distMatrix(Chebyshev()),
             knnTreeFun = BruteTree,
             rStart = 0.0, rFinal=0.1, radiusFun=expRadius(-5.0),
             epochs = 20)

Overload of trainGigaSOM for simple DataFrames and matrices. This slices the data using DistributedArrays, sends them the workers, and runs normal trainGigaSOM. Data is undistributed after the computation.

GigaSOM.trainGigaSOM — Method

trainGigaSOM(som::Som, dInfo::LoadedDataInfo;
             kernelFun::Function = gaussianKernel,
             metric = Euclidean(),
             somDistFun = distMatrix(Chebyshev()),
             knnTreeFun = BruteTree,
             rStart = 0.0, rFinal=0.1, radiusFun=expRadius(-5.0),
             epochs = 20)

Arguments:

som: object of type Som with an initialised som

dInfo: LoadedDataInfo object that describes a loaded dataset

kernelFun::function: optional distance kernel; one of (bubbleKernel, gaussianKernel) default is gaussianKernel

metric: Passed as metric argument to the KNN-tree constructor

somDistFun: Function for computing the distances in the SOM map

knnTreeFun: Constructor of the KNN-tree (e.g. from NearestNeighbors package)

rStart: optional training radius. If zero (default), it is computed from the SOM grid size.

rFinal: target radius at the last epoch, defaults to 0.1

radiusFun: Function that generates radius decay, e.g. linearRadius or expRadius(10.0)

epochs: number of SOM training iterations (default 10)

GigaSOM.bubbleKernel — Method

bubbleKernel(x, r::Float64)

Return a "bubble" (spherical) distribution kernel.

GigaSOM.distMatrix — Function

distMatrix(metric=Chebyshev())

Return a function that uses the metric (compatible with metrics from package Distances) calculates distance matrixes from normal row-wise data matrices, using the metric.

Use as a parameter of trainGigaSOM.

GigaSOM.expRadius — Function

expRadius(steepness::Float64)

Return a function to be used as a radiusFun of trainGigaSOM, which causes exponencial decay with the selected steepness.

Use: trainGigaSOM(..., radiusFun = expRadius(0.5))

Arguments

steepness: Steepness of exponential descent. Good values range from -100.0 (almost linear) to 100.0 (really quick decay).

GigaSOM.gaussianKernel — Method

gaussianKernel(x, r::Float64)

Return the value of normal distribution PDF (σ=r, μ=0) at x

GigaSOM.gridRectangular — Method

gridRectangular(xdim, ydim)

Create coordinates of all neurons on a rectangular SOM.

The return-value is an array of size (Number-of-neurons, 2) with x- and y- coordinates of the neurons in the first and second column respectively. The distance between neighbours is 1.0. The point of origin is bottom-left. The first neuron sits at (0,0).

Arguments

xdim: number of neurons in x-direction

ydim: number of neurons in y-direction

GigaSOM.linearRadius — Method

linearRadius(initRadius::Float64, iteration::Int64, decay::String, epochs::Int64)

Return a neighbourhood radius. Use as the radiusFun parameter for trainGigaSOM.

Arguments

initRadius: Initial Radius

finalRadius: Final Radius

iteration: Training iteration

epochs: Total number of epochs

GigaSOM.thresholdKernel — Function

thresholdKernel(x, r::Float64)

Simple FlowSOM-like hard-threshold kernel

Embedding

GigaSOM.embedGigaSOM — Method

embedGigaSOM(som::GigaSOM.Som,
             data;
             knnTreeFun = BruteTree,
             metric = Euclidean(),
             k::Int64=0,
             adjust::Float64=1.0,
             smooth::Float64=0.0,
             m::Float64=10.0)

Overload of embedGigaSOM for simple DataFrames and matrices. This slices the data using DistributedArrays, sends them the workers, and runs normal embedGigaSOM. All data is properly undistributed after the computation.

Examples:

Produce a 2-column matrix with 2D cell coordinates:

e = embedGigaSOM(som, data)

Plot the result using 2D histogram from Gadfly:

using Gadfly
draw(PNG("output.png",20cm,20cm),
     plot(x=e[:,1], y=e[:,2],
     Geom.histogram2d(xbincount=200, ybincount=200)))

GigaSOM.embedGigaSOM — Method

embedGigaSOM(som::GigaSOM.Som,
             dInfo::LoadedDataInfo;
             knnTreeFun = BruteTree,
             metric = Euclidean(),
             k::Int64=0,
             adjust::Float64=1.0,
             smooth::Float64=0.0,
             m::Float64=10.0,
             output::Symbol=tmpSym(dInfo))::LoadedDataInfo

Return a data frame with X,Y coordinates of EmbedSOM projection of the data.

Arguments:

som: a trained SOM

dInfo: LoadedDataInfo that describes the loaded dataset

knnTreeFun: Constructor of the KNN-tree (e.g. from NearestNeighbors package)

metric: Passed as metric argument to the KNN-tree constructor

k: number of nearest neighbors to consider (high values get quadratically slower)

adjust: position adjustment parameter (higher values avoid non-local approximations)

smooth: approximation smoothness (the higher the value, the larger the neighborhood of approximate local linearity of the projection)

m: exponential decay rate for the score when approaching the k+1-th neighbor distance

output: variable name for storing the distributed result

Data must have the same number of dimensions as the training dataset, and must be normalized using the same parameters.

GigaSOM.embedGigaSOM_internal — Method

embedGigaSOM_internal(som::GigaSOM.Som,
                      data::Matrix{Float64},
                      tree,
                      k::Int64,
                      adjust::Float64,
                      boost::Float64,
                      m::Float64)

Internal function to compute parts of the embedding on a prepared kNN-tree structure (tree) and smooth converted to boost.

Distributed processing tools

GigaSOM.distribute_array — Method

distribute_array(sym, x::Array, pids; dim=1)::LoadedDataInfo

Distribute roughly equal parts of array x separated on dimension dim among pids into a worker-local variable sym.

Returns the LoadedDataInfo structure for the distributed data.

GigaSOM.distribute_darray — Method

distribute_darray(sym, dd::DArray)::LoadedDataInfo

Distribute the distributed array parts from dd into worker-local variable sym.

Returns the LoadedDataInfo structure for the distributed data.

GigaSOM.distributed_collect — Function

distributed_collect(dInfo::LoadedDataInfo, dim=1; free=false)

Distributed collect (just as the other overload) that works with LoadedDataInfo.

GigaSOM.distributed_collect — Function

distributed_collect(val::Symbol, workers, dim=1; free=false)

Collect the arrays distributed on workers under value val into an array. The individual arrays are pasted in the dimension specified by dim, i.e. dim=1 is roughly equivalent to using vcat, and dim=2 to hcat.

val must be an Array-based type; the function will otherwise fail.

If free is true, the val is undistributed after collection.

This preallocates the array for results, and is thus more efficient than e.g. using distributed_mapreduce with vcat for folding.

GigaSOM.distributed_exec — Method

distributed_exec(val, fn, workers)

Execute a function on workers, taking val as a parameter. Results are not collected. This is optimal for various side-effect-causing computations that are not expressible with distributed_transform.

GigaSOM.distributed_exec — Method

distributed_exec(dInfo::LoadedDataInfo, fn)

Variant of distributed_exec that works with LoadedDataInfo.

GigaSOM.distributed_foreach — Method

distributed_foreach(arr::Vector, fn, workers)

Call a function fn on workers, with a single parameter arriving from the corresponding position in arr.

GigaSOM.distributed_mapreduce — Method

distributed_mapreduce(val, map, fold, workers)

Run maps (non-modifying transforms on the data) and folds (2-to-1 reductions) on the worker-local data (in vals) distributed on workers and return the final reduced result.

It is assumed that the fold operation is associative, but not commutative (as in semigroups). If there are no workers, operation returns nothing (we don't have a monoid to magically conjure zero elements :[ ).

In current version, the reduce step is a sequential left fold, executed in the main process.

Example

# compute the mean of all distributed data
sum,len = distributed_mapreduce(:myData,
    (d) -> (sum(d),length(d)),
    ((s1, l1), (s2, l2)) -> (s1+s2, l1+l2),
    workers())
println(sum/len)

Processing multiple arguments (a.k.a. "zipWith")

The val here does not necessarily need to refer to a symbol, you can easily pass in a quoted tuple, which will be unquoted in the function parameter. For example, distributed values :a and :b can be joined as such:

distributed_mapreduce(:((a,b)),
    ((a,b)::Tuple) -> [a b],
    vcat,
    workers())

GigaSOM.distributed_mapreduce — Method

distributed_mapreduce(dInfo1::LoadedDataInfo, dInfo2::LoadedDataInfo, map, fold)

Variant of distributed_mapreduce that works with more LoadedDataInfos at once. The data must be distributed on the same set of workers, in the same order.

GigaSOM.distributed_mapreduce — Method

distributed_mapreduce(vals::Vector, map, fold, workers)

Variant of distributed_mapreduce that works with more distributed variables at once.

GigaSOM.distributed_mapreduce — Method

distributed_mapreduce(dInfo::LoadedDataInfo, map, fold)

Distributed map/reduce (just as the other overload of distributed_mapreduce) that works with LoadedDataInfo.

GigaSOM.distributed_transform — Function

distributed_transform(dInfo::LoadedDataInfo, fn, tgt::Symbol=dInfo.val)::LoadedDataInfo

Same as distributed_transform, but specialized for LoadedDataInfo.

GigaSOM.distributed_transform — Function

distributed_transform(val, fn, workers, tgt::Symbol=val)

Transform the worker-local distributed data available as val on workers in-place, by a function fn. Store the result as tgt (default val)

Example

# multiply all saved data by 2
distributed_transform(:myData, (d)->(2*d), workers())

GigaSOM.get_from — Method

get_from(worker,val)

Get a value val from a remote worker; quoting of val works just as with save_at. Returns a future with the requested value.

GigaSOM.get_val_from — Method

get_val_from(worker,val)

Shortcut for instantly fetching the future from get_from.

GigaSOM.remove_from — Method

remove_from(worker,sym)

Sets symbol sym on worker to nothing, effectively freeing the data.

GigaSOM.save_at — Method

save_at(worker, sym, val)

Saves value val to symbol sym at worker. sym should be quoted (or contain a symbol). val gets unquoted in the processing and evaluated at the worker, quote it if you want to pass exact command to the worker.

This is loosely based on the package ParallelDataTransfers, but made slightly more flexible by omitting/delaying the explicit fetches etc. In particular, save_at is roughly the same as ParallelDataTransfers.sendto, and get_val_from works very much like ParallelDataTransfers.getfrom.

Return value

A future with Nothing that can be fetched to see that the operation has finished.

Examples: addprocs(1) saveat(2,:x,123) # saves 123 saveat(2,:x,myid()) # saves 1 saveat(2,:x,:(myid())) # saves 2 saveat(2,:x,:(:x)) # saves the symbol :x # (just :x won't work because of unquoting)

Note: Symbol scope

The symbols are saved in Main module on the corresponding worker. For example, save_at(1, :x, nothing) will erase your local x variable. Beware of name collisions.

GigaSOM.tmpSym — Method

tmpSym(dInfo::LoadedDataInfo; prefix="", suffix="_tmp")

Decorate the symbol from dInfo with prefix and suffix.

GigaSOM.tmpSym — Method

tmpSym(s::Symbol; prefix="", suffix="_tmp")

Decorate a symbol s with prefix and suffix, to create a good name for a related temporary value.

GigaSOM.undistribute — Method

undistribute(dInfo::LoadedDataInfo)

Remove the loaded data described by dInfo from the corresponding workers.

GigaSOM.undistribute — Method

undistribute(sym, workers)

Remove the loaded data from workers.

GigaSOM.defaultFiles — Method

defaultFiles(s, pids)

Make a good set of filenames for saving a dataset.

GigaSOM.distributed_export — Function

distributed_export(dInfo::LoadedDataInfo, files=defaultFiles(dInfo.val, dInfo.workers))

Overloaded functionality for LoadedDataInfo.

GigaSOM.distributed_export — Function

distributed_export(sym::Symbol, pids, files=defaultFiles(sym,pids))

Export the content of symbol sym by each worker specified by pids to a corresponding filename in files.

GigaSOM.distributed_import — Function

distributed_import(dInfo::LoadedDataInfo, files=defaultFiles(dInfo.val, dInfo.workers))

Overloaded functionality for LoadedDataInfo.

GigaSOM.distributed_import — Function

distributed_import(sym::Symbol, pids, files=defaultFiles(sym,pids))

Import the content of symbol sym by each worker specified by pids from the corresponding filename in files.

GigaSOM.distributed_unlink — Function

distributed_unlink(dInfo::LoadedDataInfo, files=defaultFiles(dInfo.val, dInfo.workers))

Overloaded functionality for LoadedDataInfo.

GigaSOM.distributed_unlink — Function

distributed_unlink(sym::Symbol, pids, files=defaultFiles(sym,pids))

Remove the files created by distributed_export with the same parameters.